requirements.txt
Telethon==1.38.1
tqdm==4.67.1
colorama==0.4.6
cryptg==0.5.0.post0
python-dotenv==1.0.1
click==8.1.7
rich==13.7.0
aiofiles==23.2.1
tenacity==8.2.3
PyYAML==6.0.1 